+++ lib/base_core.dart
@@ -16,6 +16,7 @@ export 'package:base_core/src/data/drivers/firebase/firebase_crashlytics_driver.
 export 'package:base_core/src/data/drivers/firebase/firebase_database_driver.dart';
 export 'package:base_core/src/data/drivers/firebase/firebase_driver.dart';
 export 'package:base_core/src/data/drivers/firebase/firebase_notifications_driver.dart';
+export 'package:base_core/src/data/drivers/firebase/firebase_remote_config_driver.dart';
 export 'package:base_core/src/data/drivers/firebase/firebase_storage_driver.dart';
 export 'package:base_core/src/data/drivers/local_auth_driver.dart';
 export 'package:base_core/src/data/drivers/local_notifications_driver.dart';
@@ -44,6 +45,7 @@ export 'package:base_core/src/domain/interfaces/either.dart';
 export 'package:base_core/src/domain/services/firebase/i_firebase_auth_service.dart';
 export 'package:base_core/src/domain/services/firebase/i_firebase_database_service.dart';
 export 'package:base_core/src/domain/services/firebase/i_firebase_notifications_service.dart';
+export 'package:base_core/src/domain/services/firebase/i_firebase_remote_service.dart';
 export 'package:base_core/src/domain/services/firebase/i_firebase_storage_service.dart';
 export 'package:base_core/src/domain/services/i_app_tracking_service.dart';
 export 'package:base_core/src/domain/services/i_connectivity_service.dart';
@@ -60,6 +62,7 @@ export 'package:base_core/src/infra/drivers/firebase/i_firebase_auth_driver.dart
 export 'package:base_core/src/infra/drivers/firebase/i_firebase_database_driver.dart';
 export 'package:base_core/src/infra/drivers/firebase/i_firebase_driver.dart';
 export 'package:base_core/src/infra/drivers/firebase/i_firebase_notifications_driver.dart';
+export 'package:base_core/src/infra/drivers/firebase/i_firebase_remote_driver.dart';
 export 'package:base_core/src/infra/drivers/firebase/i_firebase_storage_driver.dart';
 export 'package:base_core/src/infra/drivers/i_app_tracking_driver.dart';
 export 'package:base_core/src/infra/drivers/i_connectivity_driver.dart';
@@ -86,6 +89,7 @@ export 'package:base_core/src/infra/services/firebase/firebase_auth_service.dart
 export 'package:base_core/src/infra/services/firebase/firebase_crashlytics_service.dart';
 export 'package:base_core/src/infra/services/firebase/firebase_database_service.dart';
 export 'package:base_core/src/infra/services/firebase/firebase_notifications_service.dart';
+export 'package:base_core/src/infra/services/firebase/firebase_remote_service.dart';
 export 'package:base_core/src/infra/services/firebase/firebase_storage_service.dart';
 export 'package:base_core/src/infra/services/local_auth_service.dart';
 export 'package:base_core/src/infra/services/local_notifications_service.dart';