+++ lib/src/domain/services/firebase/i_firebase_remote_service.dart
@@ -1,0 +1,9 @@
+import '../../../../base_core.dart';
+
+abstract class IFirebaseRemoteService {
+  Future<Either<Exception, Object>> getAll();
+Future<Either<Exception, Object>> getDouble({required String key});
+Future<Either<Exception, Object>> getInt({required String key});
+Future<Either<Exception, Object>> getString({required String key});
+Future<Either<Exception, Object>> getValue({required String key});
+}